requirements.txt
fastapi>=0.110
uvicorn>=0.29
jinja2>=3.1
python-multipart>=0.0.9
httpx>=0.27
pydantic>=2.7
python-pptx>=0.6.23
openai>=1.33.0
anthropic>=0.30.0
google-generativeai>=0.7.2
